p2p: use size_t for arbitrary counters instead of uint8_t

+++ src/p2p/net_node.inl
@@ -1856,8 +1856,8 @@ namespace nodetool
   template<class t_payload_net_handler>
   bool node_server<t_payload_net_handler>::has_too_many_connections(const epee::net_utils::network_address &address)
   {
-    const uint8_t max_connections = 1;
-    uint8_t count = 0;
+    const size_t max_connections = 1;
+    size_t count = 0;
 
     m_net_server.get_config_object().foreach_connection([&](const p2p_connection_context& cntxt)
     {
